+++ netbox_ssl/api/urls.py
@@ -9,5 +9,6 @@ from . import views
 router = NetBoxRouter()
 router.register("certificates", views.CertificateViewSet)
 router.register("assignments", views.CertificateAssignmentViewSet)
+router.register("csrs", views.CertificateSigningRequestViewSet)
 
 urlpatterns = router.urls