+++ netbox_ssl/forms/__init__.py
@@ -8,6 +8,12 @@ from .certificates import (
     CertificateForm,
     CertificateImportForm,
 )
+from .csr import (
+    CertificateSigningRequestBulkEditForm,
+    CertificateSigningRequestFilterForm,
+    CertificateSigningRequestForm,
+    CSRImportForm,
+)
 
 __all__ = [
     "CertificateForm",
@@ -16,4 +22,8 @@ __all__ = [
     "CertificateImportForm",
     "CertificateAssignmentForm",
     "CertificateAssignmentFilterForm",
+    "CertificateSigningRequestForm",
+    "CertificateSigningRequestFilterForm",
+    "CertificateSigningRequestBulkEditForm",
+    "CSRImportForm",
 ]