+++ netbox_ssl/graphql/schema.py
@@ -5,7 +5,7 @@ GraphQL schema for NetBox SSL plugin.
 import strawberry
 import strawberry_django
 
-from .types import CertificateAssignmentType, CertificateType
+from .types import CertificateAssignmentType, CertificateSigningRequestType, CertificateType
 
 
 @strawberry.type(name="Query")
@@ -35,6 +35,18 @@ class NetBoxSSLQuery:
         from ..models import CertificateAssignment
 
         return CertificateAssignment.objects.all()
+
+    @strawberry_django.field
+    def certificate_signing_request(self, id: int) -> CertificateSigningRequestType:
+        from ..models import CertificateSigningRequest
+
+        return CertificateSigningRequest.objects.get(pk=id)
+
+    @strawberry_django.field
+    def certificate_signing_request_list(self) -> list[CertificateSigningRequestType]:
+        from ..models import CertificateSigningRequest
+
+        return CertificateSigningRequest.objects.all()
 
 
 schema = [NetBoxSSLQuery]